Add comparison with last screenshot before saving
This is taking the nuclear option, but it prevents situations where
duplicate images are saved, because the user opened the Snip & Sketch
main window after taking a screenshot, etc.

This is because, when you open the main window, the screenshot is
copied to the clipboard again. Additionally, every time the user draws
on the screenshot, the image is sent to the clipboard twice, for some
reason...

Currently, the comparison code considers the most recently created
file in the screenshot output directory to be the last saved
screenshot. This could be optimised in the future, by remembering the
last saved screenshot after we save it, to prevent having to comb the
filesystem every time.