Switch from GUID to `uID` notification area ID
So, I found out why different GUIDs were needed for debug and release
builds... It's because a notification area icon's GUID is tied to a
specific program path. Therefore, moving the executable, or running a
debug / release build, with the same GUID, caused a conflict, causing
the notification area icon creation process to fail.

This is supposedly to stop spoofing of notification area icons, but
people could just create icons with old IDs that look like the program
they're spoofing still...

As it turns out, you can use GUIDs and allow people to move the
program. All you have to do is sign your program using Authenticode...
($$$$$)

Therefore, we just use old IDs. The downside of this is that, if the
program is forcefully terminated, the icon is duplicated when it is
next started, but oh well.

Source: https://docs.microsoft.com/en-gb/windows/win32/api/shellapi/ns-shellapi-notifyicondataa?redirectedfrom=MSDN#troubleshooting